Move Hash32 and Hash20 to package primitive
