+++ genvm/templates/multisig/types.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"fmt"
 
+	// required to help scalegen find the types
+	_ "github.com/spacemeshos/go-spacemesh/common/types/primitive"
 	"github.com/spacemeshos/go-spacemesh/genvm/core"
 	"github.com/spacemeshos/go-spacemesh/genvm/templates/wallet"
 )
