+++ hash/hash.go
@@ -21,3 +21,12 @@ func Sum(chunks ...[]byte) (rst [32]byte) {
 	hh.Sum(rst[:0])
 	return rst
 }
+
+func Sum20(chunks ...[]byte) (rst [20]byte) {
+	hh := New()
+	for _, chunk := range chunks {
+		hh.Write(chunk)
+	}
+	hh.Digest().Read(rst[:])
+	return rst
+}
